+++ oj/unmashall_test.go
@@ -74,3 +74,15 @@ func TestUnmarshaler(t *testing.T) {
 	tt.Equal(t, 1, len(tags))
 	tt.Equal(t, 1, tags["k1"])
 }
+
+type Triple [3]float64
+
+func TestUnmarshalArray(t *testing.T) {
+	var tri Triple
+	src := []byte(`[1.0, 2.0, 3.0]`)
+	err := oj.Unmarshal(src, &tri)
+	tt.Nil(t, err)
+	tt.Equal(t, 1.0, tri[0])
+	tt.Equal(t, 2.0, tri[1])
+	tt.Equal(t, 3.0, tri[2])
+}