+++ CHANGELOG.md
@@ -11,7 +11,7 @@ The structure and content of this file follows [Keep a Changelog](https://keepac
   JSONPath document in section 2.7.
 - Added `jp.Expr.Locate()` function that returns normalized paths for JSONPath expression.
 ### Fixed
-- TBD Unmarshal now supports arrays such as `[4]int`.
+- Unmarshal now supports arrays such as `[4]int`.
 
 ## [1.20.3] - 2023-11-09
 ### Added